See what happens if we setinputsizes for String
Per Anthony Tuininga, setinputsizes() is important
particularly if you have a unicode value you want to line
up with an index under Python 3, because you want the type
to be bound using NVARCHAR2 and not VARCHAR2.   Only if
we setinputsizes() for NVARCHAR2 will we get that effect.
So see at least if setinputsizes() works for STRING these
days or otherwise figure out what the problem was.

Additionally, the type object now for NVARCHAR2 is
cx_Oracle.NCHAR, so use that, and also remove some
old _cx_oracle_XYZ collections we aren't using to
reduce verbosity.

Change-Id: I4977f54796485ec454b20697ec20db6d70fb6f84